src/test/javascript/spec/app/entities/m-match-formation/m-match-formation.service.spec.ts
/* tslint:disable max-line-length */
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { HttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { take, map } from 'rxjs/operators';
import { MMatchFormationService } from 'app/entities/m-match-formation/m-match-formation.service';
import { IMMatchFormation, MMatchFormation } from 'app/shared/model/m-match-formation.model';

describe('Service Tests', () => {
  describe('MMatchFormation Service', () => {
    let injector: TestBed;
    let service: MMatchFormationService;
    let httpMock: HttpTestingController;
    let elemDefault: IMMatchFormation;
    let expectedResult;
    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = {};
      injector = getTestBed();
      service = injector.get(MMatchFormationService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new MMatchFormation(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0);
    });

    describe('Service methods', () => {
      it('should find an element', async () => {
        const returnedFromService = Object.assign({}, elemDefault);
        service
          .find(123)
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: elemDefault });
      });

      it('should create a MMatchFormation', async () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );
        const expected = Object.assign({}, returnedFromService);
        service
          .create(new MMatchFormation(null))
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));
        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: expected });
      });

      it('should update a MMatchFormation', async () => {
        const returnedFromService = Object.assign(
          {
            position1: 1,
            position2: 1,
            position3: 1,
            position4: 1,
            position5: 1,
            position6: 1,
            position7: 1,
            position8: 1,
            position9: 1,
            position10: 1,
            position11: 1
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);
        service
          .update(expected)
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));
        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: expected });
      });

      it('should return a list of MMatchFormation', async () => {
        const returnedFromService = Object.assign(
          {
            position1: 1,
            position2: 1,
            position3: 1,
            position4: 1,
            position5: 1,
            position6: 1,
            position7: 1,
            position8: 1,
            position9: 1,
            position10: 1,
            position11: 1
          },
          elemDefault
        );
        const expected = Object.assign({}, returnedFromService);
        service
          .query(expected)
          .pipe(
            take(1),
            map(resp => resp.body)
          )
          .subscribe(body => (expectedResult = body));
        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a MMatchFormation', async () => {
        const rxPromise = service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
